fix(ci): close three CI-on-CI failures
Three real failures observed on the latest commit's CI run, each fixed
in place:

1. `xtask/src/ci/plan.rs` had `#[allow(clippy::too_many_arguments)]`
   declared twice on `pub fn run`, which `clippy::duplicated_attributes`
   (in `clippy::all`, fatal under `-D warnings`) caught in workspace-
   wide Clippy. Removed the duplicate.

2. `crates/bitnet-ffi/src/config.rs` had `u64::from(self.seed)` where
   `self.seed: u64`. This is a pre-existing latent clippy::useless_conversion
   that PR 07 (workspace lint inheritance) made visible: bitnet-ffi
   gained `[lints] workspace = true` and now inherits `clippy::all =
   warn`, which `RUSTFLAGS=-D warnings` in the Feature Matrix workspace
   Clippy promotes to error. Fixed inline by removing the no-op cast.

3. `policy/non-rust-allowlist.toml` did not cover `ripr.toml` (added in
   PR 13). The Policy workflow runs `check-file-policy --fail-on-error`
   and it caught the gap. Added an explicit `[[allow]]` entry with
   owner `testing/oracle`, surface `ci`, classification `config`,
   covered_by `[xtask check-file-policy, .github/workflows/ripr.yml]`.

PR Gate Success failed last run only because `Policy` failed; this
commit unblocks both.

Acceptance:
  cargo clippy --locked --workspace --no-default-features --features cpu \
    -- -D warnings                                                  -> clean
  cargo run -p xtask -- check-file-policy --fail-on-error           -> 8165 files / 113 allow / 0 findings
  cargo run -p xtask -- ci-lane-whitelist check                     -> 15 lanes / 2 exceptions / 0 errors
  cargo fmt --all -- --check                                        -> clean

+++ crates/bitnet-ffi/src/config.rs
@@ -283,7 +283,7 @@ impl BitNetCInferenceConfig {
         config.repetition_penalty = self.repetition_penalty;
 
         if self.seed != 0 {
-            config = config.with_seed(u64::from(self.seed));
+            config = config.with_seed(self.seed);
         }
 
         config
